requirements.txt
# Scan QC System - Backend Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# 3D Processing
open3d>=0.17.0
trimesh>=4.0.0
numpy>=1.24.0
scipy>=1.10.0
rtree>=1.0.0

# PDF Generation
reportlab>=4.0.0
PyMuPDF>=1.22.0  # fitz for PDF text extraction

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0  # For PCA in visualization

# AI Integration (choose one or more)
anthropic>=0.40.0    # Claude 4.5
# google-generativeai>=0.8.0  # Gemini 3
# openai>=1.50.0      # GPT-5.2

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0
